+++ delta-kernel-unitycatalog/src/commits_client/mod.rs
@@ -2,10 +2,11 @@ use serde::Deserialize;
 use tracing::instrument;
 use url::Url;
 
-use crate::config::ClientConfig;
-use crate::error::Result;
-use crate::http::{build_http_client, execute_with_retry, handle_response};
-use crate::models::commits::{CommitRequest, CommitsRequest, CommitsResponse};
+use unity_catalog_client::config::ClientConfig;
+use unity_catalog_client::error::Result;
+use unity_catalog_client::http::{build_http_client, execute_with_retry, handle_response};
+
+use crate::models::{CommitRequest, CommitsRequest, CommitsResponse};
 
 #[cfg(any(test, feature = "test-utils"))]
 mod in_memory;
@@ -16,9 +17,8 @@ pub use in_memory::{InMemoryCommitsClient, TableData};
 /// Trait for UC commits API operations.
 ///
 /// Implementations of this trait are responsible for performing any necessary retries on transient
-/// failures. This trait is designed to be injected into a `uc_catalog::committer::UCCommitter`,
-/// which itself does not perform any retries and relies on the underlying client implementation to
-/// handle retry logic.
+/// failures. This trait is designed to be injected into a [`crate::UCCommitter`], which itself does
+/// not perform any retries and relies on the underlying client implementation to handle retry logic.
 #[allow(async_fn_in_trait)]
 pub trait UCCommitsClient: Send + Sync {
     /// Get the latest commits for the table.
