+++ delta-kernel-unitycatalog/tests/data/catalog_managed_0/part1=1/part-00000-8afb1c56-2018-4af2-aa4f-4336c1b39efd.c000.snappy.parquet
