+++ unitycatalog-client/src/error.rs
