+++ unitycatalog-client/src/models/credentials.rs
