+++ unitycatalog-client/src/models/mod.rs
@@ -1,7 +1,5 @@
-pub mod commits;
 pub mod credentials;
 pub mod tables;
 
-pub use commits::{Commit, CommitRequest, CommitsRequest, CommitsResponse};
 pub use credentials::{AwsTempCredentials, TemporaryTableCredentials};
 pub use tables::TablesResponse;
